requirements.txt
packaging==23.2
PyMuPDF
langgraph==0.0.32
langchain_openai==0.1.1
langchain==0.1.13
html2text
python-docx
json5
python-dotenv
beautifulsoup4
selenium
webdriver_manager
streamlit
arxiv
praw
tinydb
fpdf
